+++ NachoSpigot-Server/src/main/java/net/minecraft/server/EntityTNTPrimed.java
@@ -30,9 +30,9 @@ public class EntityTNTPrimed extends Entity {
         this.setPosition(d0, d1, d2);
         float f = (float) (Math.random() * 3.1415927410125732D * 2.0D);
 
-        this.motX = (double) (-((float) Math.sin((double) f)) * 0.02F);
+        this.motX = -((float) Math.sin(f)) * 0.02F;
         this.motY = 0.20000000298023224D;
-        this.motZ = (double) (-((float) Math.cos((double) f)) * 0.02F);
+        this.motZ = -((float) Math.cos(f)) * 0.02F;
         this.fuseTicks = 80;
         this.lastX = d0;
         this.lastY = d1;
@@ -161,7 +161,7 @@ public class EntityTNTPrimed extends Entity {
         double d4 = this.locY + this.getHeadHeight() - d1;
         double d5 = this.locZ - d2;
 
-        return (double) MathHelper.sqrt(d3 * d3 + d4 * d4 + d5 * d5);
+        return MathHelper.sqrt(d3 * d3 + d4 * d4 + d5 * d5);
     }
 
     @Override
